Moved _related_search to prevent circular dependency
# Conflicts:
#	api/main/rest/_media_query.py
